// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        InvoiceOut: {
            /** Id */
            id: number;
            /** Po Id */
            po_id: number;
            /** Invoice Number */
            invoice_number: string | null;
            /**
             * Invoice Date
             * Format: date
             */
            invoice_date: string;
            /** Invoice Value */
            invoice_value: string;
            /** Expected Payment Date */
            expected_payment_date: string | null;
            /** Client Tds Applicable */
            client_tds_applicable: boolean;
            /** Client Tds Amount */
            client_tds_amount: string;
            /** Amount Expected */
            amount_expected: string | null;
            /**
             * Payment Status
             * @enum {string}
             */
            payment_status: "pending" | "partially_paid" | "paid";
            /** Actual Amount Received */
            actual_amount_received: string | null;
            /** Actual Receipt Date */
            actual_receipt_date: string | null;
            /** Value Override */
            value_override: boolean;
            /** Override Remarks */
            override_remarks: string | null;
            /** Invoice Pdf Url */
            invoice_pdf_url: string | null;
            /**
             * Status
             * @enum {string}
             */
            status: "draft" | "final";
            /** Allocations */
            allocations: components["schemas"]["InvAllocationOut"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}